% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{account_to_steam}
\alias{account_to_steam}
\title{Convert 32-bit account id to a 64 bit Steam id}
\usage{
account_to_steam(account_id)
}
\arguments{
\item{account_id}{The 32 bit account id, can be numeric or character.}
}
\value{
A string representing a users 64 bit steam id.
}
\description{
Converts a 32-bit account id to a 64 bit Steam id that can be used to query
api's that require the 64 bit steam id
}
\examples{
account_to_steam(32500026)
}

