% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{details}
\alias{details}
\title{Details}
\usage{
details(api_key, match_ids, delay = 0.5, simplify_colnames = TRUE, ...)
}
\arguments{
\item{api_key}{A String of your steam developer key, see: \url{https://steamcommunity.com/dev/apikey}}

\item{match_ids}{A list or a single match id that we will query the api with}

\item{delay}{The amount of time (in seconds) to wait between api calls.}

\item{simplify_colnames}{}
}
\value{
A data.frame of the match details
}
\description{
Query the api for each match_id in match_ids,
returns the results in a dataframe
}
\examples{
details(api_key = "<api_key>", match_ids = c(2331506594, 2331349045))
}

