% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{match_details_url}
\alias{match_details_url}
\title{Match Details url}
\usage{
match_details_url(api_key = api_key, match_id, https = TRUE)
}
\arguments{
\item{api_key}{A String of your steam developer key, see: \url{https://steamcommunity.com/dev/apikey}}

\item{match_id}{The id of the match that you wish to get more information about.}

\item{https}{Logical, if True: https, False: http}
}
\value{
The url for querying the dota 2 api.
}
\description{
Builds a URL for retrieving match details
}
\examples{
match_details_url(api_key = "<your_api_key>", match_id = 123114122)
}

