% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{match_history_url}
\alias{match_history_url}
\title{Match History url}
\usage{
match_history_url(api_key = api_key, https = TRUE, ...)
}
\arguments{
\item{api_key}{A String of your steam developer key, see: \url{https://steamcommunity.com/dev/apikey}}

\item{https}{Logical, if True: https, False: http}

\item{...}{Query parameters, see: \url{http://wiki.teamfortress.com/wiki/WebAPI/GetMatchHistory}}
}
\value{
The url for querying the dota 2 api.
}
\description{
Builds a URL for retrieving match histories from the dota 2 api.
}
\details{
When called with no query parameters it will generate a url for  retrieveing the last 100 public games.

see: \url{http://wiki.teamfortress.com/wiki/WebAPI/GetMatchHistory}
}
\examples{
match_history_url(api_key = "<your_api_key>")
}

