% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{raw_match_details}
\alias{raw_match_details}
\title{Match Details as an R object}
\usage{
raw_match_details(api_key, match_id, ...)
}
\arguments{
\item{api_key}{A String of your steam developer key, see: \url{https://steamcommunity.com/dev/apikey}}

\item{match_id}{The id of the match that you wish to get more information about.}
}
\description{
Visits the url generated by \code{\link{match_details_url}}
and parses the JSON into an R object
}

