% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{raw_match_history}
\alias{raw_match_history}
\title{Match history as an R object}
\usage{
raw_match_history(api_key, ...)
}
\arguments{
\item{api_key}{A String of your steam developer key, see: \url{https://steamcommunity.com/dev/apikey}}

\item{...}{Query parameters, see: \url{http://wiki.teamfortress.com/wiki/WebAPI/GetMatchHistory}}
}
\description{
Visits the url generated by \code{\link{match_history_url}}
and parses the JSON into an R object
}

