% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{url_builder}
\alias{url_builder}
\title{builds a url for the dota 2 api}
\usage{
url_builder(query, api_args, https, api = "IDOTA2Match_570",
  version = "V001")
}
\arguments{
\item{query}{A String The type of query to build GetMatchHistory or GetMatchDetails}

\item{api_args}{A list of parameters for the query}

\item{https}{Logical if True: https, False: http}

\item{api}{string: The api id}
}
\value{
the url of the query
}
\description{
See: https://wiki.teamfortress.com/wiki/WebAPI
}

