% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{all_matches}
\alias{all_matches}
\title{Get a all of a players matches}
\usage{
all_matches(api_key, account_id, delay = 0.5, verbose = TRUE, ...)
}
\arguments{
\item{api_key}{A String of your steam developer key, see: \url{https://steamcommunity.com/dev/apikey}}

\item{account_id}{}

\item{delay}{the ammount of time in seconds to wait between api calls}

\item{verbose}{#'}

\item{...}{Query parameters, see: \url{http://wiki.teamfortress.com/wiki/WebAPI/GetMatchHistory}}
}
\value{
A list of match ids that we are then able to gain futher information on
}
\description{
Get a list of game ids that are availible for a player.
The maximum number of games that steam remembers for a player is 500
}
\examples{
all_matches(api_key = "<your key>", account_id = 12314141)
}

