% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dota2api.R
\name{url_builder}
\alias{url_builder}
\title{Builds a url for the dota 2 api}
\usage{
url_builder(query, api_args, https, api = "IDOTA2Match_570",
  version = "V001")
}
\arguments{
\item{query}{A String The type of query to build GetMatchHistory or GetMatchDetails}

\item{api_args}{A list of parameters for the query.}

\item{https}{Logical, if True: https, False: http.}

\item{api}{The string id of the api.}

\item{version}{The string of the api version.}
}
\value{
The url of the query
}
\description{
See: \url{ttps://wiki.teamfortress.com/wiki/WebAPI}
}

